pyproject.toml
[tool.poetry]
name = "books-project"
version = "0.1.0"
description = ""
authors = ["fmontes00 <fernandomontesdeoca93@gmail.com>"]
readme = "README.md"
packages = [{include = "books_project"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
djangorestframework = "^3.14.0"
black = "^23.1.0"
pillow = "^9.4.0"
django-model-utils = "^4.3.1"
ipdb = "^0.13.13"
pylint = "^2.17.0"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
pytest = "^7.2.2"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
